src-python/requirements.txt
biopython>=1.81
numpy>=1.24
openai>=1.0
pyyaml>=6.0
sqlalchemy>=2.0
openpyxl>=3.1
pyinstaller>=6.0
